--- Content.Server/Drone/Components/DroneComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.Drone.Components
-{
-    [RegisterComponent]
-    public sealed class DroneComponent : Component
-    {
-        public float InteractionBlockRange = 2.15f;
-    }
-}
