+++ Content.Shared/Drone/DroneComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared.Drone
+{
+    [RegisterComponent]
+    public sealed class DroneComponent : Component
+    {}
+}
