+++ Content.Shared/Drone/SharedDroneSystem.cs
@@ -1,9 +1,26 @@
+using Content.Shared.Interaction.Events;
+using Content.Shared.Interaction.Components;
+using Content.Shared.Item;
+using Content.Shared.Tag;
 using Robust.Shared.Serialization;
 
 namespace Content.Shared.Drone
 {
-    public abstract class SharedDroneSystem : EntitySystem
+    public class SharedDroneSystem : EntitySystem
     {
+        [Dependency] private readonly TagSystem _tagSystem = default!;
+        public override void Initialize()
+        {
+            SubscribeLocalEvent<DroneComponent, InteractionAttemptEvent>(OnInteractionAttempt);
+        }
+
+        private void OnInteractionAttempt(EntityUid uid, DroneComponent component, InteractionAttemptEvent args)
+        {
+            if (args.Target != null && !HasComp<UnremoveableComponent>(args.Target)
+                && !_tagSystem.HasAnyTag(args.Target.Value, "DroneUsable", "Trash"))
+                args.Cancel();
+        }
+
         [Serializable, NetSerializable]
         public enum DroneVisuals : byte
         {
